Allow refactor window to be shown multiple times
As part of fixing this, I wanted to change the ownership of the NSWindow
reference, so it would be easier to determine when it closes, and also to
move the singleton pattern up above the UI components.

However, I think this introduces a memory leak. The NSWindow that the refactor
controller keeps a strong reference to currently MUST not release itself when
it is closed, because doing so crashes the app when our controller deallocs.