Parse target of call expression
Discovered that there was still a lot more work to handle nested call
expressions. Planning on rewriting some of this shortly, now that I've
spent an entire night poking around at call expressions and getting a feel
for how this code MIGHT be best implemented.

Basically, this needed better support for nested call expressions that are
arguments OR targets, and improving this by parsing the target initially
seemed like a big win.