Change method call parser interface
I think it is much nicer to not force the tests to know that parsing
the translation unit is a lot of work. This also makes writing a
recursive implementation a lot easier because the public interface looks
nothing like the private interface, which complicated a bit of turning
this into a recursive implementation last night.