Rewriting project file makes the plugin installable again
Completely unsure WHAT changed here, but at some point when I had moved and
renamed some of the parser classes the plugin continued to build, and the tests
still pass, but when xcode is launched the plugin is never actually notified that
the application started, and that it should attach its menu extras.

This feels a lot like a problem I've solved in the past, but I can't seem to find
any commits in this repo or any other that refer to a similar problem. My best guess
is that we are able to generate a plugin but it's missing some crucial plist key or that
the resulting binary was missing some key class that causes it to crash at runtime and
Xcode fails to notify the user in any meaningful way (either through its UI or through a
crashlog or even just a regular log message).