Rewrite declarations of methods in headers
Found out that there is a possible bug (?) in ClangKit where asking for the tokens for
a header does not seem to correctly identify the language, resulting in the tokens not
having the correct cursor kind. This makes parsing a header a lot harder.

A plausible workaround is to use a different constructor, telling it that the file is Obj-C++.