Start identifying how to pass args to clang to expand macros
We need this so that we can actually rewrite call sites of methods in Cedar specs.

In a cedar spec, after the "using namespace" statements, we often use the SPEC_BEGIN
macro, which defines a class, allowing us to start calling some functions to define our tests.

For clang, this poses a problem because it flat-out refuses to identify that code as being valid
when it's not clearly inside of a function call (like main) or a class (like Cedar specs are, once
the macros are expanded). Hence, we need to be able to find the header and framework search paths and
pass that as args to clang.