Fix objc call expression parser to handle more edge cases
I found that call expressions were no longer being rewritten because of
several assumptions in the call expression parser. Instead of finding
selectors like initWithThis: we would get initWithThis:this -- incorrectly
assuming that the argument was a selector component. This is all kinds of
crazy because clang doesn't seem to consistently mark each part of the call
expression as different type. Sadface.