Fix command line build
Turns out that xcodebuild has different behavior when building a target
than when building a scheme. I spent the better part of an hour trying
to fixup the HEADER_SEARCH_PATHS to include Blindside, since command line
xcodebuild builds complained it could not find the umbrella header, but
the real fix was to change our xcodebuild invocation to specify a scheme
rather than the (implicit) target.