Gracefully fail when encoutering 'typealias'
Unsure how we could ever handle these, given that the choice for which
type to choose is arbitrary and entirely up to the implementor.

Maybe we could just pick String or Int at random?

Either way, this is pretty disgustingly hard because sourcekit doesn't seem
to actually expose much information about this. Might check in with SourceKitten
at some point in the future and see if they're doing something naughty here.