Fakes for non-mutating protocols are equatable
For swift classes, this is pretty easy, we can just use the ===
operator.

For structs this is ... mighty more difficult. We need to know at
compile time what types of things might be equatable, and what
types we may need to compare using more involved matters.

For example, we simply cannot compare closures or functions, which
is unfortunate because these fakes allow you to replace the entire
implementation of some methods with a closure.

http://stackoverflow.com/a/25694072

For collections (e.g.: Array), we would need to know if the underlying
type is equatable, so we can use the == operator. For something like
an array of (String, String) we need to actually compare the length,
then zip the arrays together, dereference both strings in each tuple
and then compare each string individually.

Yikes.

[#7]