Add menu item for "Implement Equatable"
Found a bug where it doesn't do the right thing if the xcode buffer
has in-memory changes.

Still needs to be documented in the README, and have some thought put
towards HOW we might expose this in a command line tool.

Much props to @jfoley and @mjstallard for the inspiration, feedback and
continuous support.