Upgrade to swift 3
Unfortunately, this will not compile because the
frameworks we were using were not ready for swift 3 either. So this
still needs to bump those frameworks. Prior to doing that, we should
delete any of the old xcode plugin nonsense and remove any dependencies
that are not necessary for fake4swift.

In fact, it might be nice to just extract fake4swift into its own repository.