Update README
No longer need to pass --HEAD to homebrew since
I resolved most (all?) of the problems tonight.

Of note, we had the wrong git tag specified (vx.y.z) vs (x.y.z),
our makefile was running in parallel, instead of waiting for
each step to succeed before attempting to move on, we did not
have git submodules inflated, and we did not specify carthage
as a dependency.

Now, some of this is pretty whack. In particular, I wish that
homebrew allowed setting arbitrary formulae as a dependency,
but I can see what that is opposed to their philosophy of not
wanting builds to download more code over the network.