Do not html escape return types
This would probably benefit from a test covering this behavior
but this is really quite an annoying piece of work, and I hope that
no one ever needs to remember the difference between {{this}} and {{{that}}}
in mustache templating.

Should you find yourself reading this commit in the future, I'm sorry, and may
Glorb have mercy upon your everlasting soul.